requirements.txt
django-cors-headers==4.1.0
djangorestframework==3.14.0
djongo==1.3.6
easydict==1.9
face-recognition==1.3.0
numpy==1.17.0
opencv_python==4.2.0.34
Pillow==7.1.2
pyrebase==3.0.27
tensorboardX==2.0
torchvision==0.4.1
# pip install torch==1.5.0+cpu -f https://download.pytorch.org/whl/torch_stable.html
# torch==1.2.0
tqdm==4.31.1